feat: split flow information from query context
