+++ src/flows/components/Flow.tsx
@@ -2,20 +2,17 @@ import React from 'react'
 
 // Components
 import PipeList from 'src/flows/components/PipeList'
-import QueryProvider from 'src/flows/context/query'
 import {RefProvider} from 'src/flows/context/refs'
 import {PopupDrawer, PopupProvider} from 'src/flows/context/popup'
 
 // NOTE: requires a FlowProvider and ResultsProvider
 const Flow = () => (
-  <QueryProvider>
-    <RefProvider>
-      <PopupProvider>
-        <PipeList />
-        <PopupDrawer />
-      </PopupProvider>
-    </RefProvider>
-  </QueryProvider>
+  <RefProvider>
+    <PopupProvider>
+      <PipeList />
+      <PopupDrawer />
+    </PopupProvider>
+  </RefProvider>
 )
 
 export default Flow
